move AuthenticatedLayout stuff into its own file
