+++ app/pages/SiloUtilizationPage.tsx
@@ -16,7 +16,7 @@ import { bytesToGiB, bytesToTiB } from '@oxide/util'
 import { useIntervalPicker } from 'app/components/RefetchIntervalPicker'
 import { SiloMetric } from 'app/components/SystemMetric'
 import { useDateTimeRangePicker } from 'app/components/form'
-import { useCurrentUser } from 'app/layouts/helpers'
+import { useCurrentUser } from 'app/layouts/AuthenticatedLayout'
 
 const toListboxItem = (x: { name: string; id: string }) => ({ label: x.name, value: x.id })
